comment changes

+++ ex3/message_slot.h
@@ -4,7 +4,7 @@
 #include <linux/ioctl.h>
 
 #define MAJOR_NUM 235
-#define MSG_SLOT_CHANNEL _IOW(MAJOR_NUM, 0, unsigned int) /*is it really int?????????*/
+#define MSG_SLOT_CHANNEL _IOW(MAJOR_NUM, 0, unsigned long)
 
 #define DEVICE_RANGE_NAME "message_slot"
 #define BUF_LEN 128
